refactor: break two import cycles and move upload I/O out of utils
Three layering violations found by an architecture audit:

- server/lib/storyBible.js re-exported createBibleStore from
  services/bibleStore.js, which imports back from it — a static cycle that
  also made ~15 genuinely-pure lib/ modules transitively pull in that
  service's filesystem CRUD. Call sites now import it from its owner.
- meatspacePost.js and meatspacePostTraining.js imported each other. The
  shared streak computation and the training-log storage primitives are now
  two one-way modules both depend on. Verified with madge --circular.
- client/src/utils/fileUpload.js made live network calls from a directory
  reserved for pure helpers. Those moved to services/apiMedia.js, which
  already owned the underlying HTTP wrappers; a re-export keeps callers
  working (see #3452 to retire it).

Also imports two universe-builder constants direct from their module
instead of through the services barrel, which dragged ~40 service modules
into a pure lib module's dependency graph.
